requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
prometheus-client==0.17.1
psycopg2-binary==2.9.7
requests==2.31.0
requests-file==1.5.1
six==1.16.0
urllib3==2.0.5
Werkzeug==2.3.7
